backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
emoji
google-genai
pymongo
beautifulsoup4
pyjwt[crypto]
Pillow
feedparser
dodopayments
